third_party/WebKit/LayoutTests/editing/execCommand/selection_table_with_empty_cells.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() =>assert_selection(
  [
    '<div contenteditable>',
      '<div>',
        '^<div>',
        '<table border="1">',
          '<tr><td>1</td><td>2</td><td>3</td></tr>',
          '<tr><td></td><td></td><td></td></tr>',
        '</table>',
        '<br>',
        '</div>',
        '<div><br></div>',
        '<div><br></div>|',
      '</div>',
    '</div>'
  ],
  'InsertUnorderedList',
  [
  '<div contenteditable>',
    '<div>',
      '<ul>',
        '<li>^',
          '<table border="1">',
          '<tbody>',
            '<tr><td>1</td><td>2</td><td>3</td></tr>',
            '<tr><td></td><td></td><td></td></tr>',
          '</tbody>',
          '</table>',
        '</li>',
        '<li>|<br></li>',
        '<li><br></li>',
      '</ul>',
    '</div>',
    '<div><br><br></div>',
  '</div>'
  ]),
  'This is test for selection containing a table with empty cells');
</script>
